Select feedback items by edited-at rather than updated-at.
The edited-at timestamp is the time that the edit was made, whereas the updated-at timestamp is the time that the DynamoDB record representing the feedback edit (`FeedbackItem`) was created/updated.  To filter by the time range of the actual edits we need to use the edited-at timestamp.

What we really need is the ability to select feedback edits based on the score configuration version number, not any timestamp.  That will come soon, once the score configurations are stored in the API and not in local YAML files.